Finishing network layer

+++ app/src/main/java/com/example/projeto_integrador/common/domain/model/movies/Genre.kt
@@ -1,6 +1,6 @@
 package com.example.projeto_integrador.common.domain.model.movies
 
 data class Genre(
-    val genreId: Int,
+    val genreId: Long,
     val genreName: String
 )
